Remove labels as input for testing wrappers
All testing wrappers received the testing labels as an input, which
makes no sense.

This was because originally the SSVM needed it, and all other wrappers
included it to maintain the signature.

Since labels where removed from sparse_svm_test in df7e3da, this input
is removed now from all wrappers.